Stop testing trim with Block DVs
after kubevirt/containerized-data-importer#2001, CDI will always
preallocate such DVs, and so all the test cases for non-preallocated
fail all the time.
Remove also the preallocated test - it's a negative test, and now
we would never get the feature on Block DVs.

Another reason to do this is that these tests did a lot of unusual
things like using rook toolbox, and despite a lot of effort, they
were unstable and quarantined for a long time.

Signed-off-by: Maya Rashish <mrashish@redhat.com>