netstat: Add unit test for SR-IOV interface without guest-agent
When an SR-IOV interface is present in the VMI spec and the guest is not
running the guest-agent, the interface is appearing in the status
section with just the network name and no other data.

On the other hand, if another non SR-IOV interface exists, the result
changes and the SR-IOV interface will not appear in the report at all.

The behavior observed is tracked through:
https://github.com/kubevirt/kubevirt/issues/7050

Signed-off-by: Edward Haas <edwardh@redhat.com>