network, fqdn: Fix adding subdomain search entry flow
The default virt-launcher dnsPolicy is ClusterFirst
(unless changed via the VMI spec).
It means that k8s default search domains will be added as the first ones,
and in case the nodes have more entries, they will be appended
to the pod's search list. [1]

The current logic of adding the missing subdomain entry,
is to find the longest entry, and to add the subdomain to this entry.

In case the longest entry will be a node entry instead of the
entries that k8s added itself, the logic would fail to find
and add the subdomain to the right service domain entry.

The fix is to find the longest svc entry instead of the absolute longest entry,
which indicates it is a k8s added service entry, and to prepend this entry
with the subdomain if needed.

We search for .svc. because it means it is a service added by k8s itself.
Only the longest service entry is the one that we should add subdomain to,
so it will be of the form subdomain.namespace.svc.cluster.local
(where cluster.local is the cluster domain).
See [2] for additional info.

In case no service entry was found a subdomain won't be added.
One example is if dnsPolicy is None and no service domain included manually.

Note:
The fix affects DHCP option 119 (domain search list).
The expected unchanged behavior is that option 15 still uses the longest entry
from the domain search list.

[1] https://kubernetes.io/docs/concepts/services-networking/dns-pod-service/#pod-s-dns-policy
[2] https://kubernetes.io/docs/tasks/debug-application-cluster/debug-service/

Signed-off-by: Or Shoval <oshoval@redhat.com>