Expose a knob to tune UninstallStrategy (#1672)
Kubevirt and CDI operator were already exposing
a knob to tune the uninstall behaviour when
workloads (VirtualMachines and DataVolumes) still
exists.
The two alternatives are:
- BlockUninstallIfWorkloadsExist
- RemoveWorkloads

HCO wasn't exposing it always hardocding BlockUninstallIfWorkloadsExist
but RemoveWorkloads can also be an interesting option
for users that wants to cascading drop all
the workloads on product removal.

The defaut behaviour will remain BlockUninstallIfWorkloadsExist

In the futere we can probably implement a better mechanism
like https://github.com/kubernetes/enhancements/pull/2840
but this is still in the design phase.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>