network, cache: Introduce a base cache object
The base cache is targeted to be used by all specialized caches and
future commits will migrate to use it.

This change also uses an internal interface that includes a subset of
the filesystem methods. This is aimed to ease testing and declare a more
focus set of methods the cache uses.

Signed-off-by: Edward Haas <edwardh@redhat.com>