virtctl expose: more legible error message
One cannot expose a VM that has no label. I had to read the code to
understand that that was my problem having read "missing label
information for vm: foo". I hope that the suggested text is more legible
and more in line with other error texts in this module.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>