api: Add UnmarshalJSON to VirtualMachineInstanceSpec
Added spacial handling for the dnsConfig.nameservers field.
In order to make it backwards compatible on go 1.17,
we remove leading zeros from IPv4 addresses, so the rest of the
system won't have spacial treatment for it.

This also adds a function to remove leading zeros from IPv4 addresses.
This function is protocol agnostic and can handle IPv6 addresses
(return them as-is).

Signed-off-by: Orel Misan <omisan@redhat.com>