Remove StorageClassBlockVolume from tests/config.go
Instead of the local block device that was created we should use StorageRWOBlock,
import the image into the PVC that is backed by the storage class and use that.
So instead of manually creating the PV and PVC from the local block storage we mapped,
we just create a DV that uses that storage class and points to the cirros image.

Signed-off-by: Shelly Kagan <skagan@redhat.com>