Increase the code coverage by taking out logic from the API package (#1692)
Each time we adding fields to the HyperConverged API, the autogenerated code cause coverage reduce. We don't need to test auto generated code.

This PR takes the version logic out of the API package, so now the API is pure data package. Then it's possible to remove the unit tests from the API package, and increase the test coverage.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>