network, netstat: Centralize the domain & agent data merging
The information collected from the domain and guest-agent is processed
in order to associate the interfaces with their corresponding networks
(when possible).
This step has been done at two locations, at the netstat update method
and at the agent-poller.

This change centralizes the data merging at the netstat update method,
removing all the processing from the virt-launcher.

The netstat unit tests no longer need to explicitly set the domain
status interface name and infoSource because these are populated in the
merge.

Signed-off-by: Edward Haas <edwardh@redhat.com>