network, netstat: Reorder and clean status update steps
Keep the various steps focus on their scope (domain, guest-agent,
cache).
The reordering is mainly focus on processing first the domain data, then
the guest-agent and pod-cache ones.
At the guest-agent step, the interfaces are associated with the domain
data (the merge) and in case of SR-IOV or SLIRP interfaces, the
IP fields are overridden by the pod-cache data.

In case no guest-agent data exists, the IP fields are overridden by the
pod-cache for all interfaces.

The changes include:
- Exposing clearly when the pod cache data is used to override domain
  and/or guest-agent data.
- The domain & guest-agent data merging is now processing only the
  interfaces that have been reported by the guest-agent. Domain-only
  interfaces are no longer handled by this step.
- The `InfoSource` is directly set at each step and no longer calculated
  as part of the guest-agent & domain merge.
  As a subsequence, the `InfoSource` field is removed from `InterfaceStatus`
- When filling up the VMI status with guest-agent data, there is no longer a
  need to condition it per the interface binding. The pod cache data will
  eventually override it anyway immediately after.

Signed-off-by: Edward Haas <edwardh@redhat.com>