virt-launcher, schema: Remove the `Name` field from `InterfaceStatus`
The `Name` field is used to associate the reported guest-agent
interfaces with the network name that is specified in the domain spec.

With the move of this association/merge to netstat (used by
virt-handler), there is no longer a need to use this field.

This change removes the field, its usages and provides alternatives:
- The explicit "merge" between the guest-agent and domain data is removed, as
  this is already taking place implicitly (as part of
  `ifacesStatusFromGuestAgent`).
- The association of SR-IOV interfaces are now done directly on the VMI
  interfaces status (and not through the guest-agent report).

Signed-off-by: Edward Haas <edwardh@redhat.com>