network, netstat: Detect SR-IOV interfaces from the domain
SR-IOV devices have been reported to the VMI status based on the
guest-agent report only.

With this change, the domain hostdevices are examined to include the
SR-IOV interfaces in the VMI status report.
Even if the guest-agent is not active, SR-IOV devices will be reported.

Fix: https://github.com/kubevirt/kubevirt/issues/7050

Signed-off-by: Edward Haas <edwardh@redhat.com>