Fix hotplugs volumesnapshot was not created in vmsnapshot
The hotplugs were included in the vmsnapshotcontent
but volumesnapshot wasnt created for them since the
volumes list was taken from the vm revision and
not from the current vm volumes.
Added the vmsnapshot hotplugs test of snapshot
and restore a check for the volumebackup and volumesnapshot
and number of restores.

Fixed: BZ#2042908

Signed-off-by: Shelly Kagan <skagan@redhat.com>