Don't call shutdown callback for zombie process
There is no need to do the shutdown for a monitored zombie process since
it has already exited. Previously it could lead to calling the callback
with pid == 0 and result in sending SIGTERM to the current process
group.

During a test run, sending SIGTERM to PID 0 will interrupt the execution
with an error.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>