builder: embed the asciidoctor utility into the builder container
In the current situation the asciidcotor tool is being used from inside
gradle, as a plugin. The repo of the plugin will be shutted down soon.
Therefore we do the transition to call this tool directly from the builder
container. This will also reduce the dependeny in a third party repo.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>