vmi_controller_test: drop non-required test
The purpose of vmi.Spec.Domain.Devices.DisableHotplug is to let the VM
owner to express that they do not want hot-pluggability nor the
additional scsi controller that comes with it. There may be a rare VM
owner who cares, so this test is fine.

However, no one cares how hot plug is implemented. No test should fail
if we ever reimplement it without a scsi controller. A kubevirt
functional tests should not look at the libvirt layer unless there is a
very good reason to do so. Our typical user wants to ensure that hotplug
works and does not require seing a scsi controller there.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>