console_test: drop redundant test_id:1594
test_id:1593 tests that we cannot connect to the serial console of a VMI
that does not run. Many years ago,
https://github.com/kubevirt/kubevirt/pull/1323/commits/29a997f2474#diff-45bcf217
introduced test_id:1594 which tests the same thing but uses a test-only
function console.NewExpecter().

I am not sure that it makes sense to wait 30s for this test. But let
alone not do it twice.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>