console_test: explain test_id:1592 a bit
I fail to understand the point of this test. It creates a VM and tries
to connect to its serial console ASAP. The test aims to aprove that the
the serial console blocks until the VM is up. I hope that my added text
makes it easier for future readers to understand the motivation and the
minor race in its implementation.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>