Make migration cancel tests reliable
Instead of trying to run stress tests and trying to abort the migration
on an overloaded VMI, create a MigrationPolicy which throttles the
migration speed to 1Mi which creates a huge time window for canceling
the migration.

It works in parallel since we can limit the throtteling to the test
namesapce without changing the global configuration and makes the tests
faster because we don't have to wait for the guest to be fully booted.

Signed-off-by: Roman Mohr <rmohr@redhat.com>