Use /mnt instead of /tmp for hostpath hotplug test
On certain OSs /tmp is tmpfs, and when the hotplug code tries
to resolve the actual path on the host, it uses findmnt to do
so, and if the actual path is on tmpfs it is unable to locate
the real path. Now using hostpath on tmpfs is not something
someone will do anyway so this commit modifies the path used
from potentially a tmpfs to a real path.

Signed-off-by: Alexander Wels <awels@redhat.com>