tests, network, expose: avoid bool arg
newLabeledVMI() use to do two different things: define a new VMI, and
then optionally create it on the cluster based on a boolean argument.

A casual reader of
    newLabeledVMI("vm", virtClient, true)
cannot tell what "true" means. It is cleaner and simpler to explicitly
call tests.RunVMIAndExpectLaunch() if running the VMI is required.

This replaced code uses the equivalent of
tests.RunVMIAndExpectLaunchIgnoreWarnings() but I believe that this is
only because it is a remnant of the 2018 commit fc0d824da which
introduced expose tests. Since then KubeVirt has gone a long way, so I
assume the warnings are gone. Extenssive testing in CI have not exposed
failure in the modified code path.

If I am mistaken, the code can reintroduce
tests.RunVMIAndExpectLaunchIgnoreWarnings() very quickly.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>