Fix race condition in monitor tests
Do not try to manage the fake qemu process from goroutines as it leads
to races and makes the tests flaky.

Additionally simplify the exit condition check and add a defered
GinkgoRecover() call to all goroutines that run the monitoring loop.
That should prevent the test suite from panic'ing if an assertion fails
inside a goroutine.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>