tests, single stack ipv6: increase login timeout to cirros
The login is taking a bit longer since discovering ipv4 address fails
and it takes a ~minute until the discovery gives up.
Using cloud-init-nocloud userdata to turn off dhcp on eth0 interface
doesn't help to solve the issue since the usedata is invoked after the
networking.
cloudinit NetworkData or metadata.network-interfaces doesn't help as well,
since it is not properly working on cirros.

This is a temporary solution. A follow-up solution will be to use alpine
instead of cirros.

Signed-off-by: Alona Kaplan <alkaplan@redhat.com>