Remove the VMI informer from the Migration webhook
Ensure that the Migration webhook does not use a VMI informer to check
if a VMI exists. It leads to weird side-effects.

Informers can safe API calls, but should be used with caution in
webhooks. Especially singe API calls which add a complexity of O(1) to
webhooks should simply be executed without informers to avoid weird
inconsistent webhook responses.

Signed-off-by: Roman Mohr <rmohr@redhat.com>