tests: Remove unneeded RunSpecs
The RunSpecs and the other code lines of the function
are already done in an other file in the same suite.
Hence not needed.

Without removing it, ginkgo 2.0 gives the following error:
It looks like you are calling RunSpecs more than once. Ginkgo does not support
rerunning suites.  If you want to rerun a suite try ginkgo --repeat=N or
ginkgo --until-it-fails
Learn more at:
http://onsi.github.io/ginkgo/#repeating-spec-runs-and-managing-flaky-specs

Additional change:
The GinkgoT can't be in the top level, but should be in
a BeforeEach for example, because it runs DeferCleanup.

Ginkgo error otherwise:
You must call DeferCleanup inside a setup node (e.g. BeforeEach, BeforeSuite,
AfterAll...) or a subject node (i.e. It).  You can't call DeferCleanup at the
top-level or in a container node - use the After* family of setup nodes
instead.

Signed-off-by: Or Shoval <oshoval@redhat.com>