tests: Remove unneeded RunSpecs
The RunSpecs and the other lines are already done in other file
in the same suite.

If we don't remove it, ginkgo 2.0 gives the following error:

It looks like you are calling RunSpecs more than once. Ginkgo does not support
rerunning suites.  If you want to rerun a suite try ginkgo --repeat=N or
ginkgo --until-it-fails

Learn more at:
http://onsi.github.io/ginkgo/#repeating-spec-runs-and-managing-flaky-specs

Signed-off-by: Or Shoval <oshoval@redhat.com>