virt-api, unit-test: Adapt to ginkgo 2.0
Move GingkoT to BeforeEach, because it can't be on top level,
due to the implicit DeferCleanup.

The suites were missing a call to gomock.Controller.Finish(),
which verifies calls made to mocks.
With ginkgo v2, support in GinkgoT() was extended and now
Finish doesn't have to be called explicitly anymore.
The suites started to fail, as there were missing mock calls.
Add the missing ones.

Signed-off-by: Or Shoval <oshoval@redhat.com>