tests/network/probes: readiness: no need to test pod
For KubeVirt users, virt-launcher and its readiness is an implementation
detail. Only the VMI readiness is interesting.

The attempt to check virt-launcher readiness complicates the test code
too much. For exapmle, despite its name, assertPodNotReady() does not
limit itself to pod readiness; It also checks VMI readiness which is
already covered by the Consistently() line.

Let us drop it to keep the test succinct and clear.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>