tests/network/probes: replace assertPodNotReady, checkStatus, isVMIReady and vmiReady
assertPodNotReady, vmiReady, isVMIReady and checkStatus are all
functions refering to and expecting the status of the tested vmi. They
have delicately different scopes and semantics and they confuse the
reader quite a bit.

The single usage of assertPodNotReady is dropped as the test does not
really care about the state of virt-launcher pod. It is replaced by a
new Expect on getVMIConditions() which should report the precise
conditions in case the vmi is not ready, making it clearer to debug.

Similarly, checkStatus checked also the status of virt-launcher pod.
This is dropped and checking the VMI status using commonly used
tests.WaitForVMICondition().

vmiReady used to report whether an already-received VMI is ready or not.
It is dropped because it is no longer used.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>