Added ability to specify VNC proxy listening address
- Adds feature in issue #7254

General motivation
Our team discovered certain use-cases where external VNC clients were required for certain applications, such as Apache Guacamole.
The commit that added the --proxy-only feature locks the user choice down to the loopback adapter, so the only way to use the aforementioned tools is tunneling (which is not ideal for many, many ports/VMs).
This new feature will keep the old functionality and will not affect anyone using this utility unless they explicitly use the new --address flag.
The --address flag will allow for a user to specify a listening IP address. This feature is important for the following reasons:
- It allows for the user to open the listening connections to any interface/IP (0.0.0.0)
- It gives the user the capability to limit the listening to a specific interface
- It allows for the use of IPv6 specification
- This enables the ability to use with certain VPS external address configurations or VPN/tunnels that require a program to specify that address

Our team's motivation
We are using openSUSE Harvester to pioneer a new infrastructure for educational cyber ranges to give students a safe environment to study and practice malware, pentesting, etc.
We are required to use Apache Guacamole for reasons that are not negotiable. Apache Guacamole is running in the same server environment but not on the same server, so limiting this proxy to the loopback adapter makes this impossible.
At the time of writing this, our cyber range is working as intended now with the specified interface dedicated to the VNC traffic.
Specifying IP address vs just using 0.0.0.0 was also required for us, since we needed to make sure that traffic was not accessible from the other interface ports at all.
One last feature we need to implement is the ability to keep a connection listening after VNC client disconnect so students can connect back to their machines on-demand

Signed-off-by: tyleraharrison <tyleraharrison@gmail.com>