pool_test: move NewRandomPoolFromVMI out of utils
"utils" modules tend to accumulate miscelenious tools that are hard to
find. This commit attempts to reduce the load on utils.go by taking
NewRandomPoolFromVMI out of it and into its only user, pool_test. If a
second user pops up, the function should probably be placed in libvmi.

While moved, the function is renamed: there is nothing random or
unexpected about the VM pool that is being defined by it; except for the
unique name given to the pool. The argument list is made shorter to
support only what is currently needed and to avoid boolean args.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>