rpm: Update virtualization packages
Specifically:

        QEMU   6.0.0-33 -> 6.2.0-5
     libvirt    7.6.0-6 -> 8.0.0-2
  libguestfs   1.44.0-3 -> 1.44.0-5
     SeaBIOS   1.14.0-1 -> 1.15.0-1
       EDKII 20210527-1 -> 20210527-3

Since the freshest packages can now be installed directly from
the AppStream repository, we no longer need the to configure and
use the advancedvirt repository.

The latest libguestfs package has dropped a significant number
of dependencies: these are only used when dynamically generating
an appliance at runtime, and they're not needed in our case since
we use a pre-created appliance instead. As a result, the guestfs
container image is going to be smaller.

We have, however, to explicitly include tar (which is used by
the guestfs container's entrypoint.sh script) as well as file
(which is used internally by libguestfs, and whose absence from
its dependencies is a packaging issues that's in the process of
being addressed) now to keep things working.

We also have to include netcat in the launcher image because
one of the tests uses it.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>