Fix unit tests not passing ctrl.Finish
this tests needed some change since they were create
before snapshot code refactoring. The lock test only checks
for the start of the lock of updating the vm with snapshot in progress
thats why the update of the finalizers doesnt happen in the same test
run.
For the unfreeze test split it to two places calling unfreeze - one
after creating the volume snapshots when doing online vm snapshot with
guest agent. Two before unlocking the vm from snapshot in progress do
unlock if was with guest agent.

Signed-off-by: Shelly Kagan <skagan@redhat.com>