e2e: consolidate operator tests under single suite
virt-operator e2e test suite takes into account that some destructive
actions may be taken during a test, therefore the AfterEach and BeforeEach
Specs have the relevant validations.

there was a test related to virt-operator that ran separately as part
of sig-compute. According the to the logs observed from flakes it looks
that this test affects the subsequent tests, for example when running the
UpdateKubevirtConfigValueAndWait.

test has been moved under the operator_test suite.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>