Ignore sriovLiveMigration FG on SNO (#1819)
The default value of sriovLiveMigration is
true but the value is meaningless on SNO
cluster, let's simply ignore it:
the CRD defaulting mechanism on the
APIServer is not so flexible to allow to
specify a default based on the cluster topology.
On the other side, validating it with the webhook is overkilling
for users that simply expect to have the simplest possible
CR always accepted.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2062227

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>