flavors: Introduce VirtualMachineFlavorSpec
As outlined in the Flavor and preferences design proposal [1] the use of
embedded profiles within a CRD isn't prevalent within the k8s ecosystem
and might be confusing to users. As such this change extracts the
existing CPU attribute from the profile before dropping the concept
entirely from the flavors API.

Additionally this change introduces a new shared spec type for the
VirtualMachineFlavor CRDs. This drastically simplifies the code required
to find and apply these CRDs to a given VM.

[1] https://github.com/kubevirt/community/pull/165

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>