libvmi, factory: Increase memory for Cirros to 256M
Cirros image needs 256M to boot on ARM64 [1].

Although it is possible to check the node architecture in advance
and set the optimize value, it adds complexity to the code base.
The added amount is negligible and we do not have memory resource
constraints due to other VM images we support (e.g. Fedora).

Therefore, the memory resource is increased by 128M to 256M.

[1] https://github.com/kubevirt/kubevirt/issues/6363

Signed-off-by: Edward Haas <edwardh@redhat.com>