tests: Decrease timeout to login Cirros
The timeout was increased as a workarround to ipv6 only cluster
(the startup of a Cirros VM in case the is no dhcp server is very long).

Now the Cirros VMs were substituted with alpine the workarround can be
removed.

Note: the commit the timeout inclease logic from cirros login as well.
This logic was added there by mistake from the first place.


Signed-off-by: Alona Kaplan <alkaplan@redhat.com>