Add virt launcher client server memory dump command
The command will be called from the virt handler and
will trigger core dump in libvirt with memory only flag.
There will be always only one memory dump command at the same
time. Once the dump is complete it will push a timestamp to a chanel
to indicate the completion.

Signed-off-by: Shelly Kagan <skagan@redhat.com>