Handle memory dump complete event
Virt launcher notify client will wait for the memory dump
completion event. Once receiving the event it will trigger a callback
that will cause an event on the domain that will update the memory dump
info in the vmi including completion phase and the timestamp.
The update in the VMI is based on having the memory dump volume in
the volume statuses - already attached to the virt-launcher

Signed-off-by: Shelly Kagan <skagan@redhat.com>