VM with RunStrategyHalted now accepts manual stop request with graceperiod=0
Originally, RunStrategyHalted blocked all stop requests. Now, if a user needs to
force stop a VM that has stalled/crashed during shutdown, they can issue a
subsequent force stop command with a grace period of 0, which will be honored
and send a new stop request.

Stu suggested doing this with any request with a shorter grace period, but for
now I have just done it with a grace period of 0.

Signed-off-by: prnaraya <prnaraya@redhat.com>