Fix a race condition on object recreation (#1835)
If the cluster admin deletes one of the
HCO managed resource, HCO will try to
recreate it but (according to the delete
instant) HCO could potentially consume
a cached in-memoery reppresentation
of the object that can include a
resourceVersion. In that case HCO
was going to fail in a loop with:
"resourceVersion should not be set on objects to be created".
Let's ensure we never try to re-create an object
with a value for resourceVersion.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=2062321

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>