virt-launcher, hostdevice: Respect SR-IOV guest pciAddress and bootOrder
This fixes the ability to set the following fields on a list of SR-IOV NICs:
- pciAddress (as seen by the Guest)
- bootOrder

This also adds tests to check SR-IOV NICs with custom guest PCI addresses
and boot order in different configurations.

There was a closure created inside a loop, which captured
the iteration variable.
This caused the guest PCI address and boot order settings
of the last item in the list to be applied on all other list items.

For more information about capturing the iteration variable
in a closure, please see:
https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables

Signed-off-by: Orel Misan <omisan@redhat.com>