flavors: Introduce MemoryFlavor
This change introduces a new MemoryFlavor type encapsulating the Guest
and Hugepages attributes of the Memory type within a flavor. This is now
required by the VirtualMachineFlavorSpec type and thus both the
VirtualMachineFlavor and VirtualMachineClusterFlavor CRDs.

As with all flavor attributes this cannot be overridden by the user
within the VM, any attempt to do so will result in the VM being
rejected.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>