Fix a wrong unit test
fix the 'should patch if ClusterIp == \"\" during update' unit test in
/pkg/virt-operator/resource/apply/core_test.go

The test compared a slice to an empty string (""), expecting them to be
not equal. This condition will always be true and so the test allways
passed, regardless the content of the slice.

This PR fixes the test by expecting the slice to be empty.
The PR also fixes the style of the test description and the also test
description of the following test, to remove the `\"\"`, replace them by
more readable version.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>