tests, network, sriov: Use properties with multi iface spec
This is a manual cherry-pick of kubevirt#7501

When multiple SR-IOV interfaces are defined with properties (e.g. guest
PCI address), these should propagate to the guest.

The existing e2e test that covers multi SR-IOV interfaces is changed to
also check such properties, specifically the guest PCI address.

This change got triggered by a bug [1] which was not detected by the
existing tests.

[1] https://bugzilla.redhat.com/2070050

Signed-off-by: Orel Misan <omisan@redhat.com>