Add virt launcher client server memory dump command
The command will be called from the virt handler and
will trigger core dump in libvirt with memory only flag.
There will be always only one memory dump command at the same
time. Once the dump is complete it will save the inforamtion of
the last damp so next call if done for the same file will indicate
the completion or error of the last dump.

Signed-off-by: Shelly Kagan <skagan@redhat.com>