Handle a state machine of the memory dump in the vm part
The vm status is updated with a memory dump request with a
phase of binding. That triggers update of the vm and vmi volumes update
with memory dump volume source and the phase changes to inprogress.
In that time the memorydump is being mounted and done, once done or
failed the volume status in the vmi is updated with timestamp or error
in such case the state will be updated to unmounting which updates the
vmi volume list to be without the memory dump volume source which
unmounts the pvc. Once it is removed from the vmi volume status list
the phase will be updated to completed.

Signed-off-by: Shelly Kagan <skagan@redhat.com>