Handle the state machine in the vmi part
The vmi gets an update of add volume, added adjustments
to identify memory dump volume source as a hot plug.
The volume is add to the volume status list the same as a hotplug
do. Once it is attached to the virt launcher the phase is updated to
mounted. Once this Phase is updated the memory dump is triggered
and the phase is updated to in progress. Then the virt handler reconcile
loop will keep check for memory dump completion, once completed it will
update the phase and the timestamp of the memory dump which will be
recieved in the vm watch which will trigger unmount of the memory
dump volume source.

Signed-off-by: Shelly Kagan <skagan@redhat.com>