tests, vmi-configuration: Use libvmi for creating VMIs
Substitute tests.NewRandomVMIWithEphemeralDiskAndUserdata with
respective libvmi function to create Cirros, Alpine or Fedora
based VMI.

On VMI create, use the returned value to update the vmi
variable content (e.g. namespace, defaults).

It's also worth noting that tests.NewRandomVMIWithEphemeralDiskAndUserdata
adds a masquerade interface with pod network by default, even when the
test doesn't need it.
This change doesn't add the masquerade interface, as it seems better to
not clutter the tests with code adding masquerade interface, when it's
not required.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>