flavors: Ensure we have a device target before applying preferences
Users can optionally set a {Disk,CDRom,LUN}Target[1][2][3] when defining
their disks. Traditionally the vmi-admittor would default any disks
missing a target to the DiskTarget[1] type during creation [4] and would
later default the bus of the disk to sata [5].

As we are applying preferences to the VMI prior to creation we
now need to ensure a target is present before we can apply and persist
any related preferences to a DiskTarget, such as PreferredDiskBus.

[1] https://kubevirt.io/api-reference/master/definitions.html#_v1_disktarget
[2] https://kubevirt.io/api-reference/master/definitions.html#_v1_cdromtarget
[3] https://kubevirt.io/api-reference/master/definitions.html#_v1_luntarget
[4] https://github.com/kubevirt/kubevirt/blob/09445bf84b9213bb44c150b470ec9b88fbe788dd/staging/src/kubevirt.io/api/core/v1/defaults.go#L69-L75
[5] https://github.com/kubevirt/kubevirt/blob/09445bf84b9213bb44c150b470ec9b88fbe788dd/staging/src/kubevirt.io/api/core/v1/defaults.go#L127-L148

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>