migration: fix state change to MigrationPreparingTarget
This patch checks that the vmi.Status.MigrationState.MigrationUID
matches the migration being processed before proceeding in moving
the state of the migration from `MigrationScheduled` to
`MigrationPreparingTarget`.

The fix is needed because the `vmi.Status.MigrationState.MigrationUID`
field is updated during the target handoff (which happens in the
`MigrationScheduled` phase) and it might happen that if the migration
is re-processed immediately (because its key has been re-added in the workqueue)
after updating its `Phase` and the cluster is under moderate load, the
vmi Informer might have not caught up resulting in virt-controller
marking the migration as `Failed` because of a take-over by another
migration object even if it actually was successful and the VMI is still
running fine on the target node.

The false take-over error happens when the VMI has already been
previously migrated and as such the `vmi.Status.MigrationState.MigrationUID`
is already populated with some info about an old migration, specifically
the following check inside `updateStatus`:

```
if migration.TargetIsHandedOff() &&
		vmi.Status.MigrationState != nil &&
		vmi.Status.MigrationState.MigrationUID != migration.UID
```
will evaluate to true incorrectly marking the migration as failed.

This bug is easy to reproduce on kubevirtci by continously migrating
a number of VMs (in my test setup they were 3), on my setup this would
get triggered after ~300 migrations.

Signed-off-by: Antonio Cardace <acardace@redhat.com>