migration: add migration take-over e2e test
The migration take-over check which marks the migration
as failed when another migration object takes over a currently running VMI migration
is actually impossible to reach nevertheless this patch
adds a functional test to test that there is always at most
one migration active as this condition was not exercised by any
e2e test.

If two migration objects were created at the same time only one
would be able to enter the phase `MigrationScheduling` at any given time
as the check at the beginning of the `handleTargetPodCreation` function

```
	if c.podExpectations.AllPendingCreations() > 0 {
		c.Queue.AddAfter(key, 1*time.Second)
		return nil
	} else if controller.VMIActivePodsCount(vmi, c.podInformer) > 1 {
		log.Log.Object(migration).Infof("Waiting to schedule target pod for migration because there are already multiple pods running for vmi %s/%s", vmi.Namespace, vmi.Name)
		c.Queue.AddAfter(key, 1*time.Second)
		return nil

	}
```

will always reject the second migration preventing it to be scheduled.

There's also the `canMigrateVMI` function which is called at the
beginning of the sync loop that prevents a migration to be scheduled
until an already active one reaches the `Failed` or `Succeeded` state.

Signed-off-by: Antonio Cardace <acardace@redhat.com>