tests, network: Conformance migration test with emulation fix
Due to PR#6885, network conformance tests fail on environments using emulation.
The failure is caused by VM reboot that is performed as part of migration test flow.
The role of VM reboot in migration test is to rerun DHCP client initialization process
and to validate network connection after the migration is completed.
To resolve this issue, VM reboot has been replaced with DHCP client explicit restart
which is sufficient to initiate DHCP request to reset VM's IP

Signed-off-by: Diana Teplits <dteplits@redhat.com>