Fix test failures
The current test utils checks that serial tests are not running in
parallel, but the way ginkgo v2 implement serial is a bit different.

This commit allow running these tests in parallel, if their process id
is 1, because ginkgo moves all the serial tests to this process and then
wait until all the rest are done.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>