e2e: improve the virt-handler rate limiter test
It was seen that under high utilization the un-mounting process
can be way slower that the test timeouts, thus object deletion timeout
can be reached and the test will fail.

Replace a bootable VM with a non bootable one. That should
spare the test from spending time on mounting and un-mounting
the container-disk volumes as part of the VM state transition.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>