Tests: support the ginkgo --junit-report parameter
The current implementation of the functional tests, defines a junit
report test file. That causes some issues when trying to implement the
new way to run tests in parallel, using the Serial decorator.

This change checks if the ginkgo configurations, and will not generate
the junit partial files if the `--junit-report` parameter is set.

Then, the functests script does set the `--junit-report` parameter and
does not call to the junit merger anymore.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>