passt: Introduce passt interface type
Passt can be used as a simple replacement to slirp.
It is very similar to user mode networking but it implements a thinner
layer between guest and host, that only implements what's strictly needed
to pretend processes are running locally.
Same as slirp, it doesn't require any capabilities or privileges.

More info about passt can be found here: https://passt.top/passt/about

In the first phase passt will only be supported as a pod network.
It won't be supported on VMS with multus interfaces.

Signed-off-by: Alona Paz <alkaplan@redhat.com>