Add a new metric and a new alert for not completed installation (#1914)
* Prevent upgrade if one of the operand is not upgradeable (#1920)

If an operand upgradeable condition is false, HCO will set the operator
condition's Upgradeable condition to false.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Add new metric: HyperConvergedCRExists

The metric behaves as a bopolean value: if the it will be 1 if the
HyperConverged custom resource is exist, or 0 if not.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Add alert if the new metric is zero

If the new kubevirt_hco_hyperconverged_cr_exists metric is zero, then
trigger an alert that the installtion was not completed yet.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* wait an hour before start the new alert

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* fix code review comments

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>