builder: Don't install weak dependencies
We can significantly reduce the number of packages that end up in
the container by doing this.

None of the functionality that weak dependencies provide should
be needed for our use case, but even if that was the case it
would be better to explicitly list the relevant packages rather
than rely on weak dependencies.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>